# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  events.Event:
    properties:
      content:
        example: 'Issue body: Please update the connector README with setup steps.'
        type: string
      id:
        example: evt_a1b2c3d4e5f6
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      occurred_at:
        example: "2026-05-29T10:00:00Z"
        type: string
      schema_version:
        example: v1
        type: string
      source:
        example: github
        type: string
      source_id:
        example: https://github.com/owner/repo/issues/1
        type: string
      subject:
        example: 'Fix: update connector README'
        type: string
      trace_id:
        example: trace_a1b2c3d4e5f6
        type: string
      type:
        allOf:
        - $ref: '#/definitions/events.Type'
        example: document.ingested
    type: object
  events.Type:
    enum:
    - document.ingested
    - document.normalized
    - entity.extracted
    - identity.resolved
    - relationship.created
    - mismatch.detected
    - codex.analysis.completed
    type: string
    x-enum-varnames:
    - DocumentIngested
    - DocumentNormalized
    - EntityExtracted
    - IdentityResolved
    - RelationshipCreated
    - MismatchDetected
    - CodexAnalysisComplete
...
  request.SlackIngest:
    properties:
      provider:
        example: token
        type: string
      token:
        example: xoxb-111-222-xxxx
        type: string
      uri:
        example: slack://CHANNEL_ID
        type: string
      workspace_id:
        example: /home/user/myproject
        type: string
    type: object
...
  response.Ingest:
    properties:
      capabilities:
        items:
          type: string
        type: array
      connector:
        example: github
        type: string
      entity_count:
        type: integer
      event:
        $ref: '#/definitions/events.Event'
      event_count:
        example: 1
        type: integer
      events:
        items:
          $ref: '#/definitions/events.Event'
        type: array
      metadata:
        additionalProperties:
          type: string
        type: object
      metadata_items:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      mismatch_count:
        type: integer
      persisted_event_count:
        type: integer
      persistence_mode:
        type: string
      preview:
        example: 'Issue #1: Fix connector README — requesting updated setup steps.'
        type: string
      previews:
        items:
          type: string
        type: array
      relationship_count:
        type: integer
      workspace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Local-first pipeline API for ingesting and reasoning over engineering
    context.
  title: ContextOS API
  version: "1.0"
paths:
...
  /slack/callback:
    get:
      description: Receives the OAuth code from Slack, exchanges it for a token, and
        stores it locally.
      parameters:
      - description: Authorization code from Slack
        in: query
        name: code
        required: true
        type: string
      - description: CSRF state token
        in: query
        name: state
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML confirmation page
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Slack OAuth callback
      tags:
      - slack
  /slack/connect:
    get:
      description: Generates a CSRF state token and redirects to Slack's OAuth consent
        page.
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to Slack OAuth
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initiate Slack OAuth
      tags:
      - slack
  /slack/ingest:
    post:
      consumes:
      - application/json
      description: Fetches messages from a Slack channel by URI and returns a provenance-rich
        event.
      parameters:
      - description: Slack ingest request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SlackIngest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Ingest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ingest a Slack channel or message
      tags:
      - slack
  /slack/ingest/stream:
    post:
      consumes:
      - application/json
      description: Streams Codex CLI progress via SSE, then emits a result event.
      parameters:
      - description: Slack ingest request (provider must be codex)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SlackIngest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream Slack Codex ingest
      tags:
      - slack
  /slack/status:
    get:
      description: Returns token availability, source (env/oauth/none), workspace
        name, and OAuth readiness.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Slack connection status
      tags:
      - slack
...
swagger: "2.0"